Create the admin firebase class
